Clean up OpenAPI file
These extensions are both outdated and are set to the defaults, so going to remove them! The `x-headers` value is configuring the `Accept` header, which we already are doing so this is not necessary.

I also changed the `default` values to `example` values, since that's a more appropriate descriptor for what they are. We have since shipped far better support for request examples, so why not demo that!

I also updated one of the descriptions to bold the "Text" and "Value" words to match the description in their docs, even though I don't think we render this anywhere.